+++ apps/prairielearn/src/components/Navbar.types.ts
@@ -30,6 +30,7 @@ export const NavPageSchema = z
     'enroll',
     'request_course',
     'home',
+    'upgrade',
     'news_item',
     'news_items',
     'user_settings',