+++ apps/prairielearn/src/pages/instructorAssessmentUploads/instructorAssessmentUploads.ts
@@ -45,7 +45,7 @@ router.post(
 
     if (req.body.__action === 'upload_instance_question_scores') {
       const jobSequenceId = await uploadInstanceQuestionScores(
-        res.locals.assessment.id,
+        res.locals.assessment,
         req.file,
         res.locals.user.user_id,
         res.locals.authn_user.user_id,
@@ -65,7 +65,7 @@ router.post(
       }
 
       const jobSequenceId = await uploadSubmissions(
-        res.locals.assessment.id,
+        res.locals.assessment,
         req.file,
         res.locals.user.user_id,
         res.locals.authn_user.user_id,