+++ packages/sanitize/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/sanitize
 
+## 2.1.0
+
+### Minor Changes
+
+- 789fdfa: Add `truncate()` function
+
 ## 2.0.22
 
 ### Patch Changes